refactor: move sailors to their folder
