+++ sea-of-code/src/pages/home-page/components/sailors/SmokingSailor.tsx
