+++ sea-of-code/src/pages/home-page/components/sailors/ThinkingSailor.tsx
